pyproject.toml
[project]
name = "daytobananadeath"
version = "0.1.0"
description = "Banana ripeness classification with PyTorch Lightning"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "lightning>=2.0.0",
    "pillow>=9.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "torchmetrics>=0.11.0",
    "albumentations>=1.3.0",
    "wandb>=0.15.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "tensorboard>=2.20.0",
    "onnx>=1.15.0", # Pour l'export ONNX
    "onnxruntime>=1.16.0", # Pour l'inférence ONNX (optionnel)
    "uniqpath>=0.1.3",
    "fastapi>=0.118.0",
    "uvicorn[srandart,standard,standart]>=0.37.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "prometheus-client>=0.23.1",
    "pyyaml>=6.0.3",
    "python-multipart>=0.0.20",
    "ultralytics>=8.0.0", # Pour YOLO
    "pillow-heif>=1.1.1",
    "boto3>=1.40.46",
    "slowapi>=0.1.9",
    "redis>=6.4.0",
    "pydantic-settings>=2.11.0",
    "httpx>=0.28.1",
    "notebook>=7.4.7",
]
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
